Move `estimateSignedTxSize` specification to `Write.Tx.BalanceSpec`.
